output of this project
In this project their no code because of the iam using p10 and wifi3 module so there is no need to code 
pre-defined has already existed inn wifi3 module. and finally control by using app [led art app], by using this we don't need to change the code while change the sentences or any information, so easily we can use this model. 

iam upload this output pictures and videos.